1.5.0 Update / New features, new config options, code cleanup/reorganization
You can now disable the requirement of the radial menu and opt to use a "Press Key" option instead for vehicle repairs and cleans. You can now set as many locations as you'd like in the config, as well as customize the blip for each location. New config options were added. Useless code was removed and some existing code improved.